feat: auto-localization and user location search to Maps: US27
