+++ client/src/assets/images/chargingPoint.png
