--- client/src/types/searchApi.ts
@@ -1,9 +1,0 @@
-export type searchApi = {
-  geometry: {
-    coordinates: [number, number];
-  };
-
-  properties: {
-    label: string;
-  };
-};
