+++ server/src/routes/showStationsMap.routes.ts
@@ -4,6 +4,6 @@ const router = express.Router();
 
 // Define database station-related route
 import stationActions from "../modules/stations/stationActions";
-router.get("/api/station/", stationActions.browse);
+router.get("/api/stations", stationActions.browse);
 
 export default router;
